add dorfman_cost utility function
