+++ src/symgt/utils.py
@@ -1,6 +1,56 @@
 import numpy as np
 
 
+def dorfman_cost(prevalence: float, poolsize: int):
+    """
+    Compute the cost (expected number of tests per individual) according
+    to Dorfman's fixed-prevalence and infinite population analysis.
+
+    This function is a helper for `dorfman_pool_size` below.
+    """
+    p, m = float(prevalence), float(poolsize)
+    return 1 / m + 1 - (1 - p) ** m
+
+
+def dorfman_efficiency(prevalence: float, poolsize: int):
+    """
+    Compute the efficiency (expected number of individuals per test)
+    according to Dorfman's fixed-prevalence and infinite population analysis.
+    """
+    cost = dorfman_cost(prevalence, poolsize)
+    if cost == 0:
+        raise ValueError("dorfman_cost is 0")
+    return 1 / cost
+
+
+def dorfman_pool_size(prevalence: float, max_pool_size: int = 100) -> int:
+    """
+    Compute the optimal pool size according to Dorfman's infinite analysis
+    where `prevalence` is the population prevalence rate.
+
+    In other words, minimize `1/m + 1 - (1-prevalence)^m` with respect to
+    the pool size `m`.
+
+    This function is a helper for `dorfman_multfn` in ./algorithms.py.
+    """
+    if not (0.0 <= prevalence <= 1.0):
+        raise ValueError(f"prevalence={prevalence} must be in [0, 1]")
+
+    if not (max_pool_size > 1):
+        raise ValueError(f"max_pool_size={max_pool_size} should be > 1")
+
+    costs = [dorfman_cost(prevalence, m) for m in range(1, max_pool_size + 1)]
+
+    m = int(np.argmin(costs)) + 1  # off by one indexing
+
+    if 1 / m + 1 - (1 - prevalence) ** m > 1:
+        m = 1  # no pooling
+    if m == max_pool_size:
+        print("WARNING: m == max_pool_size; might need to increase max_pool_size")
+
+    return m
+
+
 def intpart_from_multfn(multfn: np.ndarray) -> np.ndarray:
     """
     Convert a multiplicity function to an integer partition.
