+++ tests/04_smoke_test:_utils.py
@@ -1,7 +1,9 @@
 import numpy as np
+import pytest
 
 from symgt.models import IIDModel
 from symgt.utils import (
+    dorfman_pool_size,
     intpart_from_multfn,
     U_from_q,
     ECost,
@@ -11,6 +13,21 @@ from symgt.utils import (
 )
 
 print("THIS IS SMOKE TEST 4: IT TESTS utils.py")
+
+# Test dorfman_cost
+
+# Test dorfman_pool_size
+assert dorfman_pool_size(0.01) == 11
+assert dorfman_pool_size(0.02) == 8
+assert dorfman_pool_size(0.05) == 5
+assert dorfman_pool_size(0.1) == 4
+assert dorfman_pool_size(0.15) == 3
+
+with pytest.raises(ValueError):
+    dorfman_pool_size(-0.1)
+with pytest.raises(ValueError):
+    dorfman_pool_size(0.1, max_pool_size=-1)
+
 
 # Test intpart_from_multfn
 assert np.allclose(intpart_from_multfn([0, 0, 1]), [2])
